[app_architecture] moving LaunchWorkflow to service package
